set up tests + fixtures

+++ tests/setup.js
@@ -17,7 +17,17 @@ afterAll(async () => {
   await mongoServer.stop();
 });
 
+// afterEach(async () => {
+//   const collections = mongoose.connection.collections;
+//   for (const key in collections) {
+//     await collections[key].deleteMany({});
+//   }
+// });
+
 afterEach(async () => {
+  // Ne pas nettoyer après chaque test si la variable SKIP_CLEANUP est définie
+  if (process.env.SKIP_CLEANUP === "true") return;
+
   const collections = mongoose.connection.collections;
   for (const key in collections) {
     await collections[key].deleteMany({});
